requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0
folium==0.15.1
streamlit-folium==0.15.0
plotly==5.18.0
streamlit-calendar==1.1.0
pydeck==0.8.0
requests==2.31.0
faker==22.1.0  # For generating demo data
matplotlib==3.8.2